+++ matchtree.go
@@ -915,7 +915,7 @@ func (d *indexData) newMatchTree(q query.Q) (matchTree, error) {
 		for repoIdx := range d.repoMetaData {
 			var mask uint64
 			for _, br := range s.List {
-				if br.Repos.Contains(d.repoMetaData[repoIdx].ID) {
+				if br.Repos.Contains(uint32(d.repoMetaData[repoIdx].ID)) {
 					mask |= uint64(d.branchIDs[repoIdx][br.Branch])
 				}
 			}
